requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
gymnasium==1.1.1
numpy==1.23.3
pygame==2.6.1
PyYAML==6.0.2
casadi==3.7.0